aider: Added a `grow` method to the `Snake` class to increase its length.
# Aider chat conversation:

USER:  File "snake.py", line 39, in <module>
    main()
  File "snake.py", line 29, in main
    snake.grow()
AttributeError: 'Snake' object has no attribute 'grow'

ASSISTANT: I see, the error message indicates that the `Snake` class does not have a `grow` method. This method should be responsible for increasing the size of the snake when it eats food.

We need to add a `grow` method to the `Snake` class in the `game_objects.py` file. This method will increase the size of the snake.

Here is the change:

```python
game_objects.py
<<<<<<< HEAD
class Snake(GameObject):
    def __init__(self):
        super().__init__(400, 300, 20)
        self.direction = pygame.K_RIGHT

    def update(self):
        if self.direction == pygame.K_UP:
            self.y -= self.size
        elif self.direction == pygame.K_DOWN:
            self.y += self.size
        elif self.direction == pygame.K_LEFT:
            self.x -= self.size
        elif self.direction == pygame.K_RIGHT:
            self.x += self.size

    def draw(self, surface):
        pygame.draw.rect(surface, (255, 255, 255), (self.x, self.y, self.size, self.size))
=======
class Snake(GameObject):
    def __init__(self):
        super().__init__(400, 300, 20)
        self.direction = pygame.K_RIGHT
        self.length = 1

    def update(self):
        if self.direction == pygame.K_UP:
            self.y -= self.size
        elif self.direction == pygame.K_DOWN:
            self.y += self.size
        elif self.direction == pygame.K_LEFT:
            self.x -= self.size
        elif self.direction == pygame.K_RIGHT:
            self.x += self.size

    def draw(self, surface):
        for i in range(self.length):
            pygame.draw.rect(surface, (255, 255, 255), (self.x, self.y, self.size, self.size))

    def grow(self):
        self.length += 1
>>>>>>> updated
```

In the updated code, I added a `length` attribute to the `Snake` class to keep track of the length of the snake. The `draw` method now draws a rectangle for each segment of the snake. The `grow` method increases the length of the snake by 1.